requirements.txt
telethon
pandas
requests
langdetect
python-dotenv
